pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "inbox-vault"
version = "0.1.0"
description = "Encrypted local email vault with AI enrichment, PII redaction, and agent-ready search"
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "beautifulsoup4>=4.12",
  "google-api-python-client>=2.70",
  "google-auth>=2.18",
  "google-auth-oauthlib>=1.0",
  "requests>=2.31",
  "sqlcipher3-binary>=0.5.4",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-cov>=5.0", "pre-commit>=3.7", "ruff>=0.6"]
redaction = ["scrubadub>=2.0"]
retrieval = ["lancedb>=0.9", "sentence-transformers>=3.0"]

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E402", "E501"]

[project.scripts]
inbox-vault = "inbox_vault.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
